🚚 Move BankAccount to api.classes package
Refactor the package structure by moving BankAccount class to
fr.milekat.banks.api.classes. Update import statements in
affected files to reflect the new package location.
