+++ mile-banks-api/src/main/java/fr/milekat/banks/api/classes/BankAccount.java
@@ -1,4 +1,4 @@
-package fr.milekat.banks.utils;
+package fr.milekat.banks.api.classes;
 
 import java.util.Map;
